NewFeature: Add Intial version of Event Filtering
The event filtering can be used to reject events from being simulated based
on the generated tracks. If the generated tracks matches the filtering condition
 this event will be simulated; otherwise, this event will be dropped, by
 cleaning the stack and regenerate a new event.

The filtering mechanism is copied from PandaRoot feb17p1, which will be merged
 into upstream in FairRoot someday (private communication with Tobias). Hence,
 the name of these classes have the Fair prefix.

KoaEvtFilterOnGeometry is designed to reject events that will not hit the recoil
detector. In this version, a very rough geometry check is implemented. Thus,
this class is intended to be modified in the future.